chore: refactor InstallGuideEditor to reuse our WebViewEditor abstraction
