+++ appland-install-guide/index.html
@@ -3,14 +3,13 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <title>AppLand Scenario</title>
+    <title>AppLand Installation Guide</title>
 </head>
-<body style="margin:0">
+<body>
 <div id="app"></div>
-<script src="./dist/installGuide.js"></script>
+<script src="./dist/main.js"></script>
 <script type="text/javascript">
-    AppLandWeb.mountInstallGuide();
+    AppLandWeb.mountWebview();
 </script>
 </body>
 </html>
