--- appland-install-guide/installGuide.js
@@ -1,8 +1,0 @@
-import Vue from 'vue';
-import { default as plugin } from '@appland/components'; // eslint-disable-line import/no-named-default
-
-export {
-  mountInstallGuide
-} from './installGuideView';
-
-Vue.use(plugin);
