+++ appland-install-guide/main.js
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+import { default as plugin } from '@appland/components'; // eslint-disable-line import/no-named-default
+
+export {
+  mountWebview
+} from './webview';
+
+Vue.use(plugin);
