--- appland-install-guide/messageListener.js
@@ -1,6 +1,0 @@
-export default class MessageListener {
-  constructor(callback, options) {
-    this.callback = callback;
-    this.options = options || {};
-  }
-}
