+++ plugin-core/src/main/java/appland/installGuide/InstallGuideEditor.java
@@ -14,22 +14,17 @@ import appland.settings.AppMapApplicationSettingsService;
 import appland.settings.AppMapProjectSettingsService;
 import appland.settings.AppMapSettingsListener;
 import appland.telemetry.TelemetryService;
-import com.google.gson.*;
-import com.intellij.execution.DefaultExecutionResult;
+import appland.webviews.WebviewEditor;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
 import com.intellij.execution.ExecutionException;
-import com.intellij.execution.ExecutionResult;
 import com.intellij.execution.Executor;
-import com.intellij.execution.configurations.CommandLineState;
-import com.intellij.execution.configurations.GeneralCommandLine;
 import com.intellij.execution.configurations.RunProfileState;
 import com.intellij.execution.executors.DefaultRunExecutor;
-import com.intellij.execution.process.KillableProcessHandler;
-import com.intellij.execution.process.ProcessHandler;
-import com.intellij.execution.process.ProcessTerminatedListener;
 import com.intellij.execution.runners.ExecutionEnvironment;
 import com.intellij.execution.runners.ExecutionEnvironmentBuilder;
-import com.intellij.execution.runners.ProgramRunner;
-import com.intellij.ide.BrowserUtil;
 import com.intellij.ide.ClipboardSynchronizer;
 import com.intellij.ide.actions.runAnything.execution.RunAnythingRunProfile;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
@@ -37,76 +32,42 @@ import com.intellij.openapi.actionSystem.impl.AsyncDataContext;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.ModalityState;
 import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.openapi.fileEditor.FileEditor;
-import com.intellij.openapi.fileEditor.FileEditorLocation;
 import com.intellij.openapi.fileEditor.FileEditorManager;
-import com.intellij.openapi.fileEditor.FileEditorState;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.util.Disposer;
-import com.intellij.openapi.util.UserDataHolderBase;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.terminal.TerminalExecutionConsole;
-import com.intellij.ui.jcef.JBCefBrowserBase;
-import com.intellij.ui.jcef.JBCefJSQuery;
-import com.intellij.ui.jcef.JCEFHtmlPanel;
 import com.intellij.util.Alarm;
 import com.intellij.util.SingleAlarm;
-import com.intellij.util.io.BaseDataReader;
-import com.intellij.util.io.BaseOutputReader;
-import org.cef.CefSettings;
-import org.cef.browser.CefBrowser;
-import org.cef.browser.CefFrame;
-import org.cef.handler.CefDisplayHandlerAdapter;
-import org.cef.handler.CefLoadHandler;
-import org.cef.handler.CefRequestHandlerAdapter;
-import org.cef.network.CefRequest;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import javax.swing.*;
 import java.awt.datatransfer.StringSelection;
-import java.beans.PropertyChangeListener;
-import java.io.IOException;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.Set;
 
-public class InstallGuideEditor extends UserDataHolderBase implements FileEditor {
-    private static final Logger LOG = Logger.getInstance("#appmap.installGuide");
-    private static final String READY_MESSAGE_ID = "intellij-plugin-ready";
+/**
+ * {@link WebviewEditor}, which displays the AppMap installation guide.
+ */
+public class InstallGuideEditor extends WebviewEditor<List<ProjectMetadata>> {
+    private static final @NotNull Logger LOG = Logger.getInstance(InstallGuideEditor.class);
 
-    private final Project project;
-    private final @NotNull VirtualFile file;
+    // currently displayed page
     private @NotNull InstallGuideViewPage type;
-
-    private final JCEFHtmlPanel contentPanel = new JCEFHtmlPanel(true, null, null);
-    private final JBCefJSQuery jcefBridge = JBCefJSQuery.create((JBCefBrowserBase) contentPanel);
-
-    private final AtomicBoolean navigating = new AtomicBoolean(false);
-    private final Gson gson = new GsonBuilder()
-            .registerTypeAdapter(AppMapMetadata.class, new AppMapMetadataWebAppSerializer())
-            .create();
-
     // to debounce the JS refresh of available AppMaps
     private final SingleAlarm projectRefreshAlarm = new SingleAlarm(this::refreshProjects, 500, this, Alarm.ThreadToUse.POOLED_THREAD);
-
     // to debounce the JS refresh when settings change
     private final SingleAlarm settingsRefreshAlarm = new SingleAlarm(this::refreshSettings, 500, this, Alarm.ThreadToUse.POOLED_THREAD);
 
-    public InstallGuideEditor(@NotNull Project project, @NotNull VirtualFile file, @NotNull InstallGuideViewPage type) {
-        this.project = project;
-        this.file = file;
+    public InstallGuideEditor(@NotNull Project project,
+                              @NotNull VirtualFile file,
+                              @NotNull InstallGuideViewPage type) {
+        super(project, file, Set.of("click-link", "open-file", "open-page", "view-problems", "clipboard",
+                "perform-install", "perform-auth", "generate-openapi"));
         this.type = type;
-
-        // contentPanel and jcefBridge register with the client as Disposable parent
-        Disposer.register(this, contentPanel.getJBCefClient());
-
-        setupListeners();
-        setupJCEF();
-        loadApplication();
     }
 
     public void navigateTo(@NotNull InstallGuideViewPage page) {
@@ -117,205 +78,242 @@ public class InstallGuideEditor extends UserDataHolderBase implements FileEditor
         postMessage(createPageNavigationJSON(page));
     }
 
-    public void refreshProjects() {
+    private void refreshProjects() {
         postMessage(createUpdateProjectsMessage());
     }
 
-    public void refreshSettings() {
+    private void refreshSettings() {
         postMessage(createUpdateSettingsMessage());
     }
 
     private void setupListeners() {
         IndexedFileListenerUtil.registerListeners(project, this, true, true, projectRefreshAlarm::cancelAndRequest);
 
-        ApplicationManager.getApplication().getMessageBus().connect(this).subscribe(AppMapSettingsListener.TOPIC, new AppMapSettingsListener() {
-            @Override
-            public void apiKeyChanged() {
-                settingsRefreshAlarm.cancelAndRequest();
-            }
+        ApplicationManager.getApplication().getMessageBus()
+                .connect(this)
+                .subscribe(AppMapSettingsListener.TOPIC, new AppMapSettingsListener() {
+                    @Override
+                    public void apiKeyChanged() {
+                        settingsRefreshAlarm.cancelAndRequest();
+                    }
 
-            @Override
-            public void enableFindingsChanged() {
-                settingsRefreshAlarm.cancelAndRequest();
-            }
+                    @Override
+                    public void enableFindingsChanged() {
+                        settingsRefreshAlarm.cancelAndRequest();
+                    }
 
-            @Override
-            public void createOpenApiChanged() {
-                settingsRefreshAlarm.cancelAndRequest();
-            }
+                    @Override
+                    public void createOpenApiChanged() {
+                        settingsRefreshAlarm.cancelAndRequest();
+                    }
 
-            @Override
-            public void openedAppMapChanged() {
-                settingsRefreshAlarm.cancelAndRequest();
-            }
-        });
+                    @Override
+                    public void openedAppMapChanged() {
+                        settingsRefreshAlarm.cancelAndRequest();
+                    }
+                });
     }
 
-    private void setupJCEF() {
-        // open links to https://appmap.io in the external browser
-        contentPanel.getJBCefClient().addRequestHandler(new CefRequestHandlerAdapter() {
-            @Override
-            public boolean onBeforeBrowse(CefBrowser browser, CefFrame frame, CefRequest request, boolean user_gesture, boolean is_redirect) {
-                return user_gesture && openExternalLink(request.getURL());
-            }
-        }, contentPanel.getCefBrowser());
-
-        contentPanel.setErrorPage((errorCode, errorText, failedUrl) -> {
-            if (errorCode == CefLoadHandler.ErrorCode.ERR_ABORTED && navigating.getAndSet(false)) {
-                return null;
-            }
-            return JBCefBrowserBase.ErrorPage.DEFAULT.create(errorCode, errorText, failedUrl);
-        });
-
-        contentPanel.getJBCefClient().addDisplayHandler(new CefDisplayHandlerAdapter() {
-            @Override
-            public boolean onConsoleMessage(CefBrowser browser, CefSettings.LogSeverity level, String message, String source, int line) {
-                if (READY_MESSAGE_ID.equals(message)) {
-                    onJavaScriptApplicationReady();
-                    return true;
-                }
-
-                var output = String.format("AppMap JS, %s:%d, %s", source, line, message);
-                switch (level) {
-                    case LOGSEVERITY_FATAL:
-                        LOG.error(output);
-                        return true;
-                    // logging ERROR as warning because the AppMap app always logs errors about svg image dimensions at start
-                    case LOGSEVERITY_ERROR:
-                    case LOGSEVERITY_WARNING:
-                        LOG.warn(output);
-                        return true;
-                    case LOGSEVERITY_INFO:
-                        LOG.info(output);
-                        return true;
-                    default:
-                        LOG.debug(output);
-                        return true;
-                }
-            }
-        }, contentPanel.getCefBrowser());
+    @Override
+    public @Nls(capitalization = Nls.Capitalization.Title) @NotNull String getName() {
+        return AppMapBundle.get("installGuide.editor.title");
     }
 
-    private void loadApplication() {
-        try {
-            contentPanel.loadURL(AppMapPlugin.getInstallGuideHTMLPath().toUri().toURL().toString());
-        } catch (IOException e) {
-            LOG.error(e);
+    @Override
+    protected @NotNull Path getApplicationFile() {
+        return AppMapPlugin.getInstallGuideHTMLPath();
+    }
+
+    @Override
+    protected @Nullable List<ProjectMetadata> createInitData() {
+        return findProjects();
+    }
+
+    @Override
+    protected void setupInitMessage(@Nullable List<ProjectMetadata> initData, @NotNull JsonObject payload) {
+        addBaseProperties(payload);
+        payload.add("projects", gson.toJsonTree(initData));
+        payload.add("disabledPages", new JsonArray());
+    }
+
+    @Override
+    protected void afterInit(@Nullable List<ProjectMetadata> initData) {
+        setupListeners();
+    }
+
+    @Override
+    protected @Nullable Gson createCustomizedGson() {
+        return new GsonBuilder()
+                .registerTypeAdapter(AppMapMetadata.class, new AppMapMetadataWebAppSerializer())
+                .create();
+    }
+
+    @Override
+    protected void handleMessage(@NotNull String messageId, @Nullable JsonObject message) throws Exception {
+        switch (messageId) {
+            case "click-link":
+                assert message != null;
+                handleMessageClickLink(message);
+                break;
+
+            case "open-file":
+                assert message != null;
+                handleMessageOpenFile(message);
+                break;
+
+            case "open-page":
+                assert message != null;
+                handleMessageOpenPage(message);
+                break;
+
+            case "view-problems":
+                handleMessageViewProblems();
+                break;
+
+            case "clipboard": {
+                assert message != null;
+                handleMessageClipboard(message);
+                break;
+            }
+
+            case "perform-install": {
+                assert message != null;
+                handleMessagePerformInstall(message);
+                break;
+            }
+
+            case "perform-auth": {
+                handleMessagePerformAuth();
+                break;
+            }
+
+            case "generate-openapi": {
+                assert message != null;
+                handleMessageGenerateOpenApi(message);
+                break;
+            }
+
+            default:
+                LOG.warn("Unhandled message type: " + type);
         }
     }
 
-    private void onJavaScriptApplicationReady() {
-        jcefBridge.addHandler(request -> {
-            LOG.debug("postMessage received message: " + request);
+    private @NotNull JsonObject createPageNavigationJSON(@NotNull InstallGuideViewPage page) {
+        var json = createMessageObject("page");
+        json.addProperty("page", page.getPageId());
+        return json;
+    }
 
-            try {
-                var json = gson.fromJson(request, JsonObject.class);
-                var type = json.has("type") ? json.getAsJsonPrimitive("type").getAsString() : null;
-                if (type != null) {
-                    switch (type) {
-                        case "postInitialize":
-                            // ignored
-                            break;
+    private @NotNull JsonObject createUpdateProjectsMessage() {
+        var json = createMessageObject("projects");
+        json.add("projects", gson.toJsonTree(findProjects()));
+        return json;
+    }
 
-                        case "click-link":
-                            if (json.has("uri")) {
-                                openExternalLink(json.getAsJsonPrimitive("uri").getAsString());
-                            }
-                            break;
+    private @NotNull JsonObject createUpdateSettingsMessage() {
+        // we're reusing the init properties to avoid duplicate code,
+        // the "settings" handler of installGuideView.js is only applying supported properties
+        var settings = createMessageObject("settings");
+        addBaseProperties(settings);
+        settings.add("projects", gson.toJsonTree(findProjects()));
+        return settings;
+    }
 
-                        case "open-file":
-                            ApplicationManager.getApplication().invokeLater(() -> {
-                                var path = Paths.get(json.getAsJsonPrimitive("file").getAsString());
-                                var file = LocalFileSystem.getInstance().findFileByNioFile(path);
-                                if (file != null) {
-                                    FileEditorManager.getInstance(project).openFile(file, true);
-                                }
-                            });
-                            break;
+    @NotNull
+    private List<ProjectMetadata> findProjects() {
+        return ProjectDataService.getInstance(project).getAppMapProjects();
+    }
 
-                        case "openAppmap":
-                            LOG.error("TODO");
-                            break;
+    /**
+     * Adds the basic JSON properties to the object, which are expected by the webview application.
+     *
+     * @param json {@link JsonObject} to modify
+     */
+    private void addBaseProperties(@NotNull JsonObject json) {
+        var settings = AppMapApplicationSettingsService.getInstance();
 
-                        case "open-page":
-                            var viewId = json.getAsJsonPrimitive("page").getAsString();
-                            var viewProject = gson.fromJson(json.get("project"), ProjectMetadata.class);
-                            var allProjects = findProjects();
-                            var anyInstallable = allProjects.stream().anyMatch(p -> p.getScore() >= 2);
-                            var projectLanguage = viewProject != null ? viewProject.getLanguage() : null;
-                            var projectLanguageName = projectLanguage != null && projectLanguage.getName() != null
-                                    ? projectLanguage.getName().toLowerCase()
-                                    : "";
+        json.addProperty("page", type.getPageId());
+        json.addProperty("userAuthenticated", settings.getApiKey() != null);
+        json.addProperty("analysisEnabled", settings.isAnalysisEnabled());
+        json.addProperty("findingsEnabled", settings.isEnableFindings());
+    }
 
-                            TelemetryService.getInstance().sendEvent(
-                                    "view:open",
-                                    (event) -> event
-                                            .property("appmap.view.id", viewId)
-                                            .property("appmap.project.language", projectLanguageName)
-                                            .property("appmap.project.installable", String.valueOf(viewProject != null && viewProject.getScore() >= 2))
-                                            .property("appmap.project.any_installable", String.valueOf(anyInstallable))
-                            );
-                            break;
+    private void handleMessageClickLink(@NotNull JsonObject message) {
+        if (message.has("uri")) {
+            openExternalLink(message.getAsJsonPrimitive("uri").getAsString());
+        }
+    }
 
-                        case "view-problems":
-                            ApplicationManager.getApplication().invokeLater(() -> {
-                                FindingsViewTab.activateFindingsTab(project);
-                            });
-                            break;
+    private void handleMessagePerformAuth() {
+        AppMapLoginAction.authenticate();
+        TelemetryService.getInstance().sendEvent("analysis:cta_interaction");
+    }
 
-                        case "clipboard": {
-                            // fixme send telemetry, as in VSCode?
-                            var content = json.getAsJsonPrimitive("text").getAsString();
-                            LOG.debug("Copying text to clipboard: " + content);
+    private void handleMessageGenerateOpenApi(@NotNull JsonObject message) {
+        var projectPath = message.has("projectPath") ? message.getAsJsonPrimitive("projectPath").getAsString() : null;
+        var appMapRoot = StringUtil.isNotEmpty(projectPath)
+                ? LocalFileSystem.getInstance().findFileByNioFile(Paths.get(projectPath))
+                : null;
 
-                            var target = new StringSelection(content);
-                            ClipboardSynchronizer.getInstance().setContent(target, target);
-                            break;
-                        }
-
-                        case "perform-install": {
-                            var path = json.getAsJsonPrimitive("path").getAsString();
-                            var language = json.getAsJsonPrimitive("language").getAsString();
-                            executeInstallCommand(path, language);
-                            break;
-                        }
-
-                        case "perform-auth": {
-                            AppMapLoginAction.authenticate();
-                            TelemetryService.getInstance().sendEvent("analysis:cta_interaction");
-                            break;
-                        }
-
-                        case "generate-openapi": {
-                            var projectPath = json.has("projectPath") ? json.getAsJsonPrimitive("projectPath").getAsString() : null;
-                            var appMapRoot = StringUtil.isNotEmpty(projectPath)
-                                    ? LocalFileSystem.getInstance().findFileByNioFile(Paths.get(projectPath))
-                                    : null;
-
-                            ApplicationManager.getApplication().invokeLater(() -> {
-                                if (appMapRoot != null) {
-                                    GenerateOpenApiAction.createOpenApiFile(project, appMapRoot, false);
-                                } else {
-                                    GenerateOpenApiAction.createOpenApiFileInteractive(project, false);
-                                }
-                            }, ModalityState.defaultModalityState());
-                            break;
-                        }
-
-                        default:
-                            LOG.warn("Unhandled message type: " + type);
-                    }
-                }
-            } catch (Exception e) {
-                LOG.warn("error handling command: " + request, e);
+        ApplicationManager.getApplication().invokeLater(() -> {
+            if (appMapRoot != null) {
+                GenerateOpenApiAction.createOpenApiFile(project, appMapRoot, false);
+            } else {
+                GenerateOpenApiAction.createOpenApiFileInteractive(project, false);
             }
+        }, ModalityState.defaultModalityState());
+    }
 
-            return new JBCefJSQuery.Response("Received " + request);
+    private void handleMessagePerformInstall(@NotNull JsonObject message) {
+        var path = message.getAsJsonPrimitive("path").getAsString();
+        var language = message.getAsJsonPrimitive("language").getAsString();
+        executeInstallCommand(path, language);
+    }
+
+    private void handleMessageClipboard(@NotNull JsonObject message) {
+        // fixme send telemetry, as in VSCode?
+        var content = message.getAsJsonPrimitive("text").getAsString();
+        LOG.debug("Copying text to clipboard: " + content);
+
+        var target = new StringSelection(content);
+        ClipboardSynchronizer.getInstance().setContent(target, target);
+    }
+
+    private void handleMessageViewProblems() {
+        ApplicationManager.getApplication().invokeLater(() -> {
+            FindingsViewTab.activateFindingsTab(project);
         });
+    }
 
-        contentPanel.getCefBrowser().executeJavaScript(createCallbackJS(jcefBridge, "postMessage"), "", 0);
+    private void handleMessageOpenFile(@NotNull JsonObject message) {
+        ApplicationManager.getApplication().invokeLater(() -> {
+            var path = Paths.get(message.getAsJsonPrimitive("file").getAsString());
+            var file = LocalFileSystem.getInstance().findFileByNioFile(path);
+            if (file != null) {
+                FileEditorManager.getInstance(project).openFile(file, true);
+            }
+        });
+    }
 
-        postMessage(createInitMessageJSON());
+    private void handleMessageOpenPage(@NotNull JsonObject message) {
+        var allProjects = findProjects();
+
+        var viewId = message.getAsJsonPrimitive("page").getAsString();
+        var viewProject = gson.fromJson(message.get("project"), ProjectMetadata.class);
+        var anyInstallable = allProjects.stream().anyMatch(p -> p.getScore() >= 2);
+        var projectLanguage = viewProject != null ? viewProject.getLanguage() : null;
+        var projectLanguageName = projectLanguage != null && projectLanguage.getName() != null
+                ? projectLanguage.getName().toLowerCase()
+                : "";
+
+        TelemetryService.getInstance().sendEvent(
+                "view:open",
+                (event) -> event
+                        .property("appmap.view.id", viewId)
+                        .property("appmap.project.language", projectLanguageName)
+                        .property("appmap.project.installable", String.valueOf(viewProject != null && viewProject.getScore() >= 2))
+                        .property("appmap.project.any_installable", String.valueOf(anyInstallable))
+        );
     }
 
     private void executeInstallCommand(String path, String language) {
@@ -350,176 +348,5 @@ public class InstallGuideEditor extends UserDataHolderBase implements FileEditor
                 LOG.warn("Failed to execute command: " + commandLine.getCommandLineString(), e);
             }
         });
-    }
-
-    private void postMessage(@NotNull JsonElement json) {
-        contentPanel.getCefBrowser().executeJavaScript("window.postMessage(" + gson.toJson(json) + ")", "", 0);
-    }
-
-    private @NotNull JsonObject createInitMessageJSON() {
-        // this contains all necessary data for the components, calculated under progress
-        var projects = findProjects();
-
-        var json = createBasePropertiesMessage("init");
-        json.add("projects", gson.toJsonTree(projects));
-        json.add("disabledPages", new JsonArray());
-        return json;
-    }
-
-    private @NotNull JsonObject createBasePropertiesMessage(@NotNull String messageType) {
-        var settings = AppMapApplicationSettingsService.getInstance();
-        var isAuthenticated = settings.getApiKey() != null;
-
-        var json = new JsonObject();
-        json.addProperty("type", messageType);
-        json.addProperty("page", type.getPageId());
-        json.addProperty("userAuthenticated", isAuthenticated);
-        json.addProperty("analysisEnabled", settings.isAnalysisEnabled());
-        json.addProperty("findingsEnabled", settings.isEnableFindings());
-        return json;
-    }
-
-    private @NotNull JsonObject createPageNavigationJSON(@NotNull InstallGuideViewPage page) {
-        var json = new JsonObject();
-        json.addProperty("type", "page");
-        json.addProperty("page", page.getPageId());
-        return json;
-    }
-
-    private @NotNull JsonObject createUpdateProjectsMessage() {
-        var json = new JsonObject();
-        json.addProperty("type", "projects");
-        json.add("projects", gson.toJsonTree(findProjects()));
-        return json;
-    }
-
-    private @NotNull JsonObject createUpdateSettingsMessage() {
-        // we're reusing the init properties to avoid duplicate code,
-        // the "settings" handler of installGuideView.js is only applying supported properties
-        var settings = createBasePropertiesMessage("settings");
-        settings.add("projects", gson.toJsonTree(findProjects()));
-        return settings;
-    }
-
-    @NotNull
-    private List<ProjectMetadata> findProjects() {
-        return ProjectDataService.getInstance(project).getAppMapProjects();
-    }
-
-    @NotNull
-    private String createCallbackJS(JBCefJSQuery query, @NotNull String functionName) {
-        return "if (!window.AppLand) window.AppLand={}; window.AppLand." + functionName + "=function(name) {" +
-                query.inject("name") + "};";
-    }
-
-    @Override
-    public @NotNull JComponent getComponent() {
-        return contentPanel.getComponent();
-    }
-
-    @Override
-    public @Nullable JComponent getPreferredFocusedComponent() {
-        return null;
-    }
-
-    @Override
-    public @Nls(capitalization = Nls.Capitalization.Title) @NotNull String getName() {
-        return AppMapBundle.get("installGuide.editor.title");
-    }
-
-    @Override
-    public void setState(@NotNull FileEditorState state) {
-    }
-
-    @Override
-    public boolean isModified() {
-        return false;
-    }
-
-    @Override
-    public boolean isValid() {
-        return true;
-    }
-
-    @Override
-    public void addPropertyChangeListener(@NotNull PropertyChangeListener listener) {
-    }
-
-    @Override
-    public void removePropertyChangeListener(@NotNull PropertyChangeListener listener) {
-    }
-
-    @Override
-    public @Nullable FileEditorLocation getCurrentLocation() {
-        return null;
-    }
-
-    @Override
-    public void dispose() {
-        LOG.debug("Disposing AppLand install guide editor");
-    }
-
-    @Override
-    public @NotNull VirtualFile getFile() {
-        return file;
-    }
-
-    public void openDevTools() {
-        ApplicationManager.getApplication().invokeLater(contentPanel::openDevtools);
-    }
-
-    private boolean openExternalLink(@Nullable String url) {
-        if (url != null && (url.startsWith("http://") || url.startsWith("https://"))) {
-            navigating.set(true);
-            BrowserUtil.browse(url);
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * RunProfileState, which is suitable for a terminal console with a PTY.
-     * The default RunAnythingRunProfileState is not suitable, but we're following it for best practices.
-     */
-    private static class AppMapRunProfileState extends CommandLineState {
-        private final GeneralCommandLine commandLine;
-
-        protected AppMapRunProfileState(@NotNull GeneralCommandLine commandLine, @NotNull ExecutionEnvironment environment) {
-            super(environment);
-            this.commandLine = commandLine;
-        }
-
-        @Override
-        public @NotNull ExecutionResult execute(@NotNull Executor executor, @NotNull ProgramRunner<?> runner) throws ExecutionException {
-            var processHandler = startProcess();
-            ProcessTerminatedListener.attach(processHandler);
-
-            var console = new TerminalExecutionConsole(getEnvironment().getProject(), processHandler);
-            console.attachToProcess(processHandler);
-            return new DefaultExecutionResult(console, processHandler, createActions(console, processHandler, executor));
-        }
-
-        @NotNull
-        @Override
-        protected ProcessHandler startProcess() throws ExecutionException {
-            // not a com.intellij.execution.process.KillableColoredProcessHandler,
-            // because the console is handling colors
-            return new KillableProcessHandler(commandLine) {
-                @Override
-                protected BaseOutputReader.@NotNull Options readerOptions() {
-                    return new BaseOutputReader.Options() {
-                        @Override
-                        public boolean splitToLines() {
-                            return false;
-                        }
-
-                        @Override
-                        public BaseDataReader.SleepingPolicy policy() {
-                            return BaseDataReader.SleepingPolicy.BLOCKING;
-                        }
-                    };
-                }
-            };
-        }
     }
 }
